controller client

+++ src/main/java/com/meriame/dao/VersementRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.meriame.model.Versement;
 
-public interface VersementRepository extends JpaRepository<Versement, Long> {
+public interface VersementRepository extends JpaRepository<Versement,  Integer> {
 
 }
